requirements.txt
beautifulsoup4==4.11.1
cleantext==1.1.4
mysql_connector_repackaged==0.3.1
requests==2.28.1
